Harden BazelPhasesDataProvider
Previously, any Bazel profile that did not include all Bazel phases
threw an InvalidProfileException. The new handling only requires
the Bazel profile to have the launch and finishing markers. If some
phases are not mentioned, it is assumed they were not executed at all.

Note that it is possible for Bazel profiles not to include all phases
and `bazel analyze-profile` successfully outputs stats on the phases.