Harden EstimatedCoresDataProvider
Previously, the estimated cores depended on specific phase markers
and events being present.
For EstimatedCoresAvailable, the previous behavior is the same, if the
phase markers are present in the profile. If they are not present, the
filters are loosened to allow still providing some data, although
potentially more error-prone.
For EstimatedCoresUsed, the impementation was changed to be more similar
to that of EstimatedCoresAvailable, as well as more flexible in that
specific events do not have to be present, and thus it's easier to loosen
the filters if some events are missing.