Fix incorrect identification of remote execution usage
Fixes #6

Previously any Bazel profile with an event of category
"Remote execution upload time" was classified as using remote execution.
However, such events are also reported when running bazel with
`--disk_cache`. Instead focus on other events, which do not seem to be reported
for invocations with `--disk_cache` that do not use remote execution.

Do not rate events with category