Fix incorrect identification of RemoteExecutionUsed (#16)
Fixes #6
Progress on #17

Previously any Bazel profile with an event of category "Remote execution
upload time" was classified as using remote execution. However, such
events are also reported when running bazel with `--disk_cache`. Instead
focus on other events, which do not seem to be reported for invocations
with `--disk_cache` that do not use remote execution.