Harden EstimatedCoresDataProvider (#11)
Fixes #10.
Progress on #17 

Previously, the estimated cores depended on specific phase markers and
events being present. Else the provider threw an error and
`EstimatedCoresUsed` and/or `EstimatedCoresAvailable` were not provided.

With this change:
* For `EstimatedCoresAvailable`, the previous behavior is the same, if
the phase markers are present in the profile. If they are not present,
the filters are loosened to allow still providing some data, although
potentially more error-prone.
* For `EstimatedCoresUsed`, the implementation was changed to be more
similar to that of `EstimatedCoresAvailable`, as well as more flexible
in that specific events do not have to be present, and thus it's easier
to loosen the filters if some events are missing.

Additionally, this change extends the tests and no longer read a profile
from a file, but uses `WriteBazelProfile` instead.