Detect queuing within bottlenecks (#30)
Currently we suggest breaking down bottleneck actions to potentially get
rid of them. However, if the bottleneck is caused by queuing, breaking
down the actions does not help.
This change additionally checks for queuing events within the bottleneck
and suggests increasing the remote execution cluster size, if the
bottleneck has significant queuing.

Additionally, this change
* adds more documentation, both in form of JavaDoc and comments, and
renaming of fields and methods
* introduces a `PartialCompleteEvent` to track an events involvement in
a bottleneck more closely
* excludes bottlenecks where the average action count is close to the
maximum number of cores used for this invocation (90% or above).

Example output:
```
Suggestion: "Avoid bottlenecks due to queuing"
These actions are involved in a bottleneck preventing parallelization:
	- abc (partially: 25m 27s of 32m 44s)
	- def (partially: 25m 27s of 33m 42s)
	- ghi (partially: 25m 27s of 26m 54s)
	- jkl (partially: 25m 25s of 30m 16s)
	- mno (partially: 24m 56s of 28m 53s)
The bottleneck includes significant queuing.
Investigate whether your remote execution cluster is overloaded. If so, consider increasing the number of workers to avoid queuing and review the cluster's autoscaling settings.
Potential improvement
The duration of the invocation can potentially be reduced by 56.88%.
[......................XXXXXXXXXXXXXXXXXXXXXXXXXXXX]
The initial build time was 42m 31s and could be reduced to 18m 20s.
Rationale
The profile includes a bottleneck lasting 25m 27s with an average action count of 9.63.
This bottleneck includes queuing for up to 25m 27s, which is 100.00% of the bottleneck's duration.
Caveats
Only the 5 longest actions out of 565 were returned. Try running this tool with --verbose to get more data.

Suggestion: "Break down bottleneck actions"
These actions are involved in a bottleneck preventing parallelization:
	- qwe (2m 53s)
	- rty (partially: 2m 43s of 4m 43s)
	- asd (partially: 2m 28s of 2m 28s)
	- fgh (2m 9s)
	- zxc (partially: 2m 7s of 4m 8s)
Try breaking them down into smaller actions.
Potential improvement
The duration of the invocation can potentially be reduced by 12.70%.
[............................................XXXXXX]
The initial build time was 42m 31s and could be reduced to 37m 7s.
Rationale
The profile includes a bottleneck lasting 5m 54s with an average action count of 16.21.
Caveats
Only the 5 longest actions out of 259 were returned. Try running this tool with --verbose to get more data.
This bottleneck includes queuing for up to 1m 8s, which is 19.32% of the bottleneck's duration.
```