Do not return EstimatedCoresAvailable, EstimatedCoresUsed, EstimatedCoresAvailable on missing data (#31)
Followup to #11.
Fixes #10.

If the relevant phase markers are not present, rather than loosening the
contraints, return a null `Datum`. The problem with loosening the
constraints is that the data is much more error-prone and this will
negatively influence the quality of the suggestions.
Instead, `DataProvider`s and `SuggestionProvider`s that need these
`Datum`s have to decide whether a reasonable evaluation is possible with
out the inputs. This change adjusts all providers currently using these
datums accordingly, as well as adds tests.