Analyzer: Add a SuggestionProvider for profiles with little data (#53)
Many of the `SuggestionProvider`s only supply suggestions if the Bazel
profile was generated for a build-like command. This change adds a
suggestion to analyze such a profile if the passed in profile looks like
it was generated for a non-build-like command (e.g. query).

Example output:
```text
Suggestion: "Analyze a different Bazel profile"
Consider analyzing a different Bazel profile. Many evaluations depend on the profile reporting some action execution. This will usually be the case when invoking a build-like Bazel command, such as `bazel build` or `bazel test`.
Rationale
The Bazel profile does not include a critical path.
The Bazel profile does not include the data required to estimate how many cores were available on the machine that the invocation was run on.
The Bazel profile does not include the data required to estimate how many cores Bazel used during the invocation.
```

Signed-off-by: Sara Adams <sara.e.adams@googlemail.com>