RemoteExecutionUsed: Avoid false positives with stricter checks
Currently th tool is too lenient with how it checks whether remote
execution was likely used. This leads to false positives, in particular
when using remote caching or a disk cache.
Instead of checking for events with the category "Remote execution process wall time",
which also occurs for non-RE builds, now check for events with the category
"remote action execution" and name "execute remotely".

Progress on #60

Signed-off-by: Sara Adams <sara.e.adams@googlemail.com>