[BazelProfile] More robust GC thread identification
The thread in which GC events are listed has been renamed multiple
times, so the code for extracting the right thread by name is
error-prone.
Instead of testing the name, search for the GC thread by its included
events. This assumes that there is only one thread that contains GC events.

Contributes to #110

Signed-off-by: Sara Adams <sara.e.adams@gmail.com>