[Skymeld] Adjust `DataProvider`s and `SuggestProvider`s for Skymeld
This change adds support for analysing profiles generated when using Skymeld.
For this, the interlaced analysis and execution phase is split into two parts:

1. Analysis only, no execution interlaced.
2. Execution has started. analysis and execution may be interlaced.

This is done by checking for the first observed `action processing` event.
The distinction is relevant, because it helps us determine how many cores
are available on the machine that ran the Bazel client, as well has how many
cores are used during execution. The later may be higher than the prior,
e.g. when remote execution is used.

* Update `SkymeldUsed` to include the combined analysis and execution phase,
    as well as optionally the phase at which execution likely starts.
* Update `EstimatedCoresDataProvider` to estimate cores when Skymeld is used.
* Update `CriticalPathNotDominantSuggestionProvider` when Skymeld is used,
    if the information when execution started is available.

Contributes to #97.

Signed-off-by: Sara Adams <sara.e.adams@gmail.com>