[DataProvider] Add provider for metrics on caching status and execution location (#138)
This data provider scans all actions and splits them into:

* remote cache hit
* remote cache miss
* remote cache not checked

as well as (for non-cache-hits):

* executed locally
* executed remotely
* execution location not reported

"Internal" Bazel actions are included in "remote cache not checked" and
"execution location not reported".
While I invested ample time in trying to single internal actions out
reliably, I did not succeed. I'm not sure it's possible with the
information currently written to profiles. A TODO to look into
separating out "internal" actions is in the code.

Contributes to #90

---------

Signed-off-by: Sara Adams <sara.e.adams@gmail.com>