[Performance] Use interner to de-duplicate strings
When reading the profile, many strings are encountered multiple
times. Use `String.interner()` to reduce the memory requirements
of running the analyzer.

Benchmark with a Bazel profile of unzipped size 357MB
Using YourKit as a profiler and checking the shallow size just before generating
the suggestions:

Before: 692 MB
After:  349 MB

Signed-off-by: Sara Adams <sara.e.adams@gmail.com>