[Core] Avoid NPEs on creating events
When creating trace event format events, the code used to throw
NullPointerExceptions when trying to construct an event from a
JsonObject that was missing required members.

This change adds static `fromJson` methods, which check the members exist,
or otherwise throw an IllegalArgumentException.
It also adds some basic tests for the Event classes.

Signed-off-by: Sara Adams <sara.e.adams@gmail.com>